requirements.txt
beautifulsoup4==4.12.3
PyYAML==6.0.2
playwright==1.48.0
